moved constants to /shared/constants.jsx, merged header and footer language selection
