+++ package.json
@@ -19,10 +19,11 @@
         "axios": "^1.7.7",
         "dompurify": "^3.1.6",
         "formik": "^2.4.6",
-        "i18next": "^23.14.0",
+        "i18next": "^23.15.1",
+        "i18next-browser-languagedetector": "^8.0.0",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
-        "react-i18next": "^15.0.1",
+        "react-i18next": "^15.0.2",
         "react-router-dom": "^6.26.1",
         "uuid": "^10.0.0",
         "yup": "^1.4.0",
