+++ src/api/auth.js
@@ -1,9 +1,5 @@
 import vaasiApiClient from './apiClient';
 
-export const INVALID_USER = 'Invalid user';
-export const USER_EMAIL_EXIST = 'Email exists';
-export const USER_PHONE_EXIST = 'Phone exists';
-
 export function authUser(body) {
     return vaasiApiClient.post(`/auth/user/signup`, body);
 }
