+++ src/shared/constants.js
@@ -2,3 +2,11 @@ export const USER_TYPE = {
     COMPANY: 'company',
     PERSON: 'person'
 };
+
+export const INVALID_USER = 'Invalid user';
+export const USER_EMAIL_EXIST = 'Email exists';
+export const USER_PHONE_EXIST = 'Phone exists';
+
+export const TOKEN_KEY = 'token';
+export const USER_ID = 'userId';
+export const PERSON_TYPE = 'personType';
